Restructure server lobby
